+++ src/steps/index.ts
@@ -3,6 +3,7 @@ import { deviceSteps } from './device';
 import { appSteps } from './app';
 import { userSteps } from './user';
 import { customProfileSteps } from './profile';
+import { blueprintSteps } from './blueprints';
 
 const integrationSteps = [
   ...accountSteps,
@@ -10,6 +11,7 @@ const integrationSteps = [
   ...deviceSteps,
   ...userSteps,
   ...customProfileSteps,
+  ...blueprintSteps,
 ];
 
 export { integrationSteps };